refactor: more cleaner arch
