+++ src/oauth/domain.rs
@@ -11,9 +11,3 @@ pub trait OAuthService: Send + Sync {
     async fn get_auth_url(&self) -> AppResult<String>;
     async fn handle_callback(&self, code: &str) -> AppResult<bool>;
 }
-
-#[async_trait]
-pub trait OAuthApiClient: Send + Sync {
-    async fn get_oauth_url(&self) -> AppResult<String>;
-    async fn exchange_code(&self, code: &str) -> AppResult<bool>;
-}
