+++ src/oauth/mod.rs
@@ -1,3 +1,2 @@
 pub mod api;
 pub mod domain;
-pub mod infra;
