--- src/shared_infra/mod.rs
@@ -1,1 +1,0 @@
-pub mod sqlite_db;
