--- src/shared_infra/sqlite_db.rs
@@ -1,16 +1,0 @@
-use sqlx::SqlitePool;
-
-#[derive(Clone)]
-pub struct SqliteDb {
-    pool: SqlitePool,
-}
-
-impl SqliteDb {
-    pub fn new(pool: SqlitePool) -> Self {
-        Self { pool }
-    }
-
-    pub fn pool(&self) -> &SqlitePool {
-        &self.pool
-    }
-}
